+++ src/provinsi/provinsi.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { accessibleBy } from '@casl/prisma';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
+import { REQUEST } from '@nestjs/core';
+import { JwtService } from '@nestjs/jwt';
+import { CaslAbilityFactory } from 'src/casl/casl-ability.factory';
 import { CreateProvinsiDto } from 'src/generated/nestjs-dto/create-provinsi.dto';
 import { UpdateProvinsiDto } from 'src/generated/nestjs-dto/update-provinsi.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -6,10 +10,21 @@ import { PrismaService } from 'src/prisma/prisma.service';
 @Injectable()
 export class ProvinsiService {
   constructor(
-    private prisma: PrismaService
+    private prisma: PrismaService,
+    private jwtService: JwtService,
+    private caslAbilityFactory: CaslAbilityFactory,
+    @Inject(REQUEST) private request: Request
   ) { }
 
   async create(createProvinsiDto: CreateProvinsiDto) {
+    const injectedToken = this.request.headers['authorization'].split(' ')[1];
+    const payload = this.jwtService.decode(injectedToken);
+    const ability = this.caslAbilityFactory.createForUser(payload);
+
+    if (!ability.can('create', 'Provinsi')) {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk membuat provinsi');
+    }
+
     const provinsi = await this.prisma.provinsi.create({
       data: {
         nama: createProvinsiDto.nama,
@@ -25,8 +40,17 @@ export class ProvinsiService {
   }
 
   async findAllProvinsiBy(params) {
+    const injectedToken = this.request.headers['authorization'].split(' ')[1];
+    const payload = this.jwtService.decode(injectedToken);
+    const ability = this.caslAbilityFactory.createForUser(payload);
+
+    if (!ability.can('read', 'Provinsi')) {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk melihat provinsi');
+    }
+
     const provinsis = await this.prisma.provinsi.findMany({
       where: {
+        AND: [accessibleBy(ability).Provinsi],
         nama: {
           contains: params.nama
         },
@@ -44,6 +68,23 @@ export class ProvinsiService {
   }
 
   async update(provinsiId: number, updateProvinsiDto: UpdateProvinsiDto) {
+    const injectedToken = this.request.headers['authorization'].split(' ')[1];
+    const payload = this.jwtService.decode(injectedToken);
+    const ability = this.caslAbilityFactory.createForUser(payload);
+
+    if (!ability.can('update', 'Provinsi')) {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk mengupdate provinsi');
+    }
+
+    await this.prisma.provinsi.findFirstOrThrow({
+      where: {
+        AND: [accessibleBy(ability).Provinsi],
+        provinsiId: provinsiId
+      }
+    }).catch(() => {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk mengupdate provinsi');
+    });
+
     const provinsi = await this.prisma.provinsi.update({
       where: {
         provinsiId
@@ -62,6 +103,23 @@ export class ProvinsiService {
   }
 
   async remove(provinsiId: number) {
+    const injectedToken = this.request.headers['authorization'].split(' ')[1];
+    const payload = this.jwtService.decode(injectedToken);
+    const ability = this.caslAbilityFactory.createForUser(payload);
+
+    if (!ability.can('delete', 'Provinsi')) {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk menghapus provinsi');
+    }
+
+    await this.prisma.provinsi.findFirstOrThrow({
+      where: {
+        AND: [accessibleBy(ability).Provinsi],
+        provinsiId: provinsiId
+      }
+    }).catch(() => {
+      throw new ForbiddenException('Anda tidak memiliki izin untuk menghapus provinsi');
+    });
+
     const provinsi = await this.prisma.provinsi.delete({
       where: {
         provinsiId
